requirements.txt
streamlit
opencv-python-headless
mahotas
joblib
numpy
scikit-learn
Pillow